Split schema into specific packages.
